Modify core.current.stubs dist dependency based on release flag
Currently, core.current.stubs.exportable.from-source output jar file is
copied to the dist directory regardless of the value of the
RELEASE_HIDDEN_API_EXPORTABLE_STUBS release flag. However, to align with
the other artifacts in the dist directory that respects the value of
the flag, the artifact of the everything stubs should be copied to the
dist directory when the flag value is not true, instead of the artifact
of the exportable stubs.

Test: lunch sdk-trunk_staging-eng && m sdk dist && inspect ninja graph
Change-Id: Id230fce9320d2250099c179b02d1103c96c9561d