Move sdk_library submodule build rules to sdk_library_internal.go
This change moves the build rules of the submodules of the
source/prebuilt java_sdk_library and the definition of the sdkLibraryXml
module type to the newly introducing sdk_library_internal.go file.

This allows sdk_library.go to solely focus on the build rules of the
sdk_library and sdk_library_import, and let sdk_library_internal.go file
to define the properties of the submodules.

This change also adds documentation of all submodules generated by the
source/prebuilt java_sdk_library.

Test: m nothing --no-skip-soong-test
Bug: 308827369
Change-Id: If3c926d257d43bc7b059c15f8854b95f428a982a