Only propgate strict updatability linting to modules that are in the apex
apex_strict_updatability_lint was propagating strict updatability
linting much further than necessary by following all transitive
dependencies, not just static dependencies of modules that are in
the apex.  This caused unnecessary lint checks, requiring the
libcore exception in b/208656169 and the allowlist in b/215736885.
Use apex.WalkPayloadDeps instead, which tracks whether the dependency
is inside the same apex or not, and stop propagating the flag when
a module outside the apex is reached.

Bug: 208656169
Bug: 215736885
Test: m lint-check
Flag: EXEMPT refactor
Change-Id: I8d801ff1414e0bdb5997d29ab453488d3ab76d64