Add AddReverseVariationDependency
Reverse variations currently must exactly match the variants of the
two modules, or only have 1 variant for the depending module. This
"1-variant fallback" is problematic for the new "base configuration"
feature and incremental soong, so we'd like to remove it. In order
to remove it, add AddReverseVariationDependency so that modules
can specify the correct variants to use instead of relying on the
1-variant fallback.

Bug: 369916167
Test: m nothing --no-skip-soong-tests
Change-Id: I2806021743de1f44041e7c818f7b84dba1aae9b3