Add PostApex mutator stage
The sabi mutator needs to run after the apex mutator because it accesses
the apex providers.  Right now it ensures it runs later by running in
the FinalDeps stage, but that prevents it from being converted to a
TransitionMutator.  Add a PostApex mutator stage that runs after PostDeps
but before FinalDeps.

Test: all soong tests pass
Bug: 367784740
Flag: EXEMPT refactor
Change-Id: I539f38ab6faa49972ec908cd54e5c80a8185ec6c