Introduce android_device module type
android_device module works as an entry point for building a specific
device image. The module can specify the name of the partition modules
that it depend on. Note that the board-specific configurations are not
specified in this module type. This will be specified in the
"configuration" modules, and the android_device module can specify the
name of the "configuration" module that it will inherit the board
configurations from.

Test: m nothing
Bug: 372486444
Change-Id: Iaaa065adc2f1018f68a33132e83caa8a6f168f80