Remove multiproduct_kati and build_test.bash
This hasn't been running in CI for over a year, and when it was
it was too slow and too impossible to keep green.  It contained
some good optimizations to check multiple configs, but with
incremental analysis we have an opportunity to rethink how to
accomplish the same goals with less resources.

Test: treehugger
Bug: 374816306
Change-Id: Id19d9c117d7474c3a5ef6ba011ada0c6b4d1dadc