Distinguish HideFromMake from SkipInstall
Invoking HideFromMake today would also make the module uninstallable.
This was a safe assumption since the packaging sytem was in build/make -
if a module is not visible to make, it is also uninstallable.

With the advent of soong packaging system, this might not always be
true. This CL disentangles SkipInstall from HideFromMake. This CL also
sets SkipInstall on modules where the installation rules should not be
emitted, e.g. to prevent duplicate installation rules between platform
and apex variants of libraries

Test: m nothing --no-skip-soong-tests
Test: no diff in aosp_cf_x86_64_phone's system file_list.txt
Change-Id: I80cdd60d2ebdba22fd37e748eb00242cc412bda1