Add NeverFar() option for transition mutators
Far variation dependencies ignore all of the source module's variations
when adding a dependency. However, there are some mutators, like the
base config mutator, that want to ensure their dependencies always
stay within the same variant. Add the ability to tag those mutators
as NeverFar mutators, so that when adding a far dependency, those
variants will still be included in the requirements for the dependency.

Bug: 361816274
Test: m nothing --no-skip-soong-tests
Change-Id: I5900d603513ccf1a4938d164c02775d650875e9b