Add boilerplates for test_suite module type to tradefed_modules
This CL adds a new module type: test_suite.
test_suite {
  name: "example-test-suite",
  description: "some example test suite"
  tests: [
    "ExampleTest1",
    "ExampleTest2",
  ],
}

test_suite allows users to define a suite of tests that can be run together. The test_suite module type takes a list of tests as input, and it will generate a single test module that can be run by tradefed.

This CL also adds a test for the test_suite module type. The test verifies that the test_suite module type can be used to define a suite of tests, and that the generated test module can be run by tradefed.

Change-Id: I1a8585899b29b0e96234a7abbd367de033d5f686
Bug: 372945132