Remove some uses of ApexContents
In preparation for removing ApexContents in order to simplify the
Soong logic and remove the top down apex mutator, delete
ApexContents.DirectlyInApex and ApexContents.InApex.

apexBundle depVisitor doesn't need to check DirectlyInApex, it can
always follow the stubs path here, and later in buildManifest the
provideNativeLibs list is removed from requireNativeLibs.

In checkStaticLinkingToStubLibraries the apex can visit direct
dependnecies to find the included shared libraries instead of
checking the ApexContents.

Bug: 372543712
Test: all apex tests pass
Test: builds
Change-Id: Ibdad6bec9f1a810058b52a4375b5e61cdc034efb