rust: Propagate static libs correctly
A dylib link is a final link, so we should not propagate staticlibs
through them. Most of the time this made no difference, but this
was breaking edge cases on coverage builds.

This CL breaks apart linkObjects so we can more finely tune which
dependencies we propagate and which we should not.

rlibs should only propagate non-whole static librares, while rlibs and
dylibs propagate whole static libs.

Bug: 377932175
Test: m
Test: New Soong tests
Change-Id: Ib2ae3dcb7d775a57bd4b1715ec528f48cc0ca026