Make ModifyPackagingSpec generic to any filesystem
ModifyPackagingSpec was added to support installing product-partition
modules on the system partition for GSI, but ramdisk also has a very
similar issue, where it's modules are installed to the "ramdisk/system"
partition.

Make ModifyPackagingSpec generic so that it works on the ramdisk files.

This reduces the ramdisk file diff down to just a missing build.prop
file.

Bug: 378146476
Test: m soong_generated_ramdisk_filesystem_test on aosp_cf_x86_64_phone
Change-Id: I1d197679a1128400b6bb837d99e3aabe2ab3c68e