Convert absolute source path to relative path in PRODUCT_COPY_FILES processing
Source paths should be relative to the top in most builds, but this is
not true for some downstream builds. Kati does not restrict this, but
Soong enforces that all source paths are relative and leads to error for
these downstream branches. This change deals with such issue by
converting the absolute paths to relative paths.

Test: m nothing --no-skip-soong-tests
Bug: 378298832
Change-Id: I313b3827417c9672b38737e8fe846bb7521b17d1