Replace InAnyApex with checking apex_available
The InAnyApex property is set by the apex transition mutator when
any apex variants are being created.  This is effectively communicating
the existence of the apex variant to the platform variant, which
complicates incremental analysis by making the platform variant
implicitly depend on the apex variant.

Replace the InAnyApex property with checking if apex_available has any
entries that are not "//apex_available:platform".

Bug: 372543712
Test: all soong tests pass
Change-Id: I2e0ae467cec0ee683ed4815a2b7397ef41375de2