Specify dtb_prebuilt property in the autogen vendor_boot.img
prebuilt_etc_modules_gen currently elides entries that are not installed
within partition directories but are installed in the product root.
However, dtb.img is copied from source via PRODUCT_COPY_FILES and
installed in the product root and then used for vendor_boot.img
generation.

This change creates the filegroup module from the PRODUCT_COPY_FILES for
the source dtb.img to utilize it in the vendor_boot.img generation, but
not actually install the file. We may consider installing this file if
this is essential in runtime.

Test: inspect the ninja command of the vendor_boot.img
Bug: 379945468
Change-Id: I760b8eb9c15b74b124828ea041ca106f8355efd4