Remove unassigned event log tags feature
.logtags files are a mapping of an integer -> name. However, there
is an unused feature where you could put a '?' instead of an integer,
and then the build system would assign a number to it, taking care
to avoid all other already-assigned numbers across the source tree.

This feature makes it hard to support some things we're working towards,
like incremental soong and treble common system images, so remove it.

Also modernize the python scripts a little.

Bug: 382515940
Test: m event-log-tags
Change-Id: Iae49e9924ac70f1ed0c141e21bc21c5d1974ef84