Support exported_symbols_list for Darwin builds
unexported_symbols_list and other linker properties are already
supported by Soong. This adds exported_symbols_list to that list.
This is needed by layoutlib_jni and libandroid_runtime to be able
to limit the symbols they export when built for host. This provides
a way to handle the dependency on a file listing the symbols to
export, similar to version_script for linux builds.


Bug: 382649667
Test: N/A
Change-Id: I1807d26506d142c0b795c59415767a6bc70a75d6