Add a `prebuilt_kernel_modules` prop to skip debug symbol strip
`Strip_debug_symbols` will default to true. If this property is set to
false explicitly, the .ko files will be installed without running
llvm-strip. The use case for this is (system|vendor|odm)_dlkm
filesystems, which package the source *.ko files without stripping
debug symbols.

This CL also fixes a bug where the output of `addLeadingSlashToPaths`
was getting ignored, since it assigned its output to a variable that was
scoped locally (via :=) instead of updating the variable in its outer
scope (via =)

Bug: 383357338
Test: mounted system_dlkm, vendor_dlkm, odm_dlkm image files built by
make and soong (top of CL stack). NOTICE is the last diff.

Change-Id: Ie5ecef0c3f6d024fe413106e17f22dee3fb18d61