Make soong's event-log-tags file match make
Make included all event-log-tags in the source tree, wether or not
they were installed on the device. Soong was only including files from
the deps of the system image.

In the long run, we'll want a per-partition event-log-tags file
for incremental soong and an independant treble system image. But
this works for now.

Bug: 382515940
Test: diff out/target/product/vsoc_x86_64/system/etc/event-log-tags out/soong/.intermediates/build/soong/fsgen/aosp_cf_x86_64_phone_generated_system_image/android_common/root/system/etc/event-log-tags
Change-Id: Id631c4292dda9531df3e460ef736fd9b15f6b566