Use {version:level} instead of % in --android-jar-pattern
Historically, Metalava only supported using `%` as a placeholder for an
API level in an `--android-jar-pattern`. Recent changes, that are part
of the work, to improve pattern matching support to allow generating
API history files from signature files have added support for using
`{version:level}` as well. This change switches the build over to using
the new placeholder, in preparation for removing support for `%`.

Bug: 383288863
Test: # Before applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
      meld before after
Change-Id: Ie75efb9d2271ef4d7a49505293cb6d113c81b0a8