Change far dependency to regular in android_device
Far variation dependencies have an undesireable trait: they pick the
closest matching variant of a module if the variants don't match and
are unspecified. This means you could end up getting a wildly different
variant of the module if you don't specify any variants like we do here.

Instead, use a regular dependency which requires that the variants
match exactly. Some of the dependencies are multilib-first, so that
requires making the android_device module also multilib-first to match.

Test: m nothing
Change-Id: I16d9b36cf8fe401cf9034a8855430c399ce356d4