apex: use proper partition tag for validations
module.PartitionTag(DeviceConfig) doesn't work for validations because
it tells the physical partition where the apex is installed. It returns
"system" even for vendor APEXes when "VendorPath" is set to
"system/vendor".

For validations, we need logical partition tag. So, "vendor:true" apexes
should be "vendor" even when it's installed in /system/vendor.

Bug: 384988749
Test: m --no-skip-soong-tests
Change-Id: Ib478caa62c98808bc94fe350a2f182f1d2b42a46