Replace --sdk-extension-root with --android-jar-pattern
Previously, the only way to specify the location of the SDK extension
jars was to specify the `--sdk-extension-root` directory and then rely
on Metalava's hard-coded pattern to find the jar files. However, it is
now possible to specify the SDK extension jar pattern explicitly as an
`--android-jar-pattern` option.

This change makes use of that new capability to specify patterns for
the SDK extension jars appropriate for the API surface being generated
and removes the `--sdk-extension-root` option altogether.

Bug: 259115852
Test: # Before applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
      meld before after
Change-Id: I4db9d4aa7b1ca3ef35a0778889f70acba491908d