Allow java_sdk_library to override the droidstubs api_surface per scope
Previously, every `java_sdk_library` would create a `droidstubs` module
passing in the name of the `apiScope` as the `api_surface` property.
e.g. the `public` scope was always called `public`, etc. While that
works for the vast majority of `java_sdk_library` modules there are a
few in `libcore`, `external/conscrypt` and `external/icu` which create
`public` scope APIs but which are not actually `public`. e.g. libcore
creates `intra-core` APIs, conscrypt and icu also create
`core-platform` APIs. They are different from `public` in the
annotations that they pass to Metalava.

A future change will pass the `api_surface` property down to Metalava
and that will be used to retrieve a corresponding configuration from
`build/soong/java/metalava/main-config.xml`. That configuration will
include the annotations. Therefore, each set of annotations that are
passed will need a unique name.

This change allows the `api_surface` for the default `public` scope
to be overridden to select the correct `api_surface` in the
configuration.

Bug: 391554590
Test: m checkapi
Change-Id: Iaa2c8325f0ffae2daf77201abe4e77c919528ad3