Configure Metalava with API surfaces used by Android
Adds basic definitions of the API surfaces used by Android to the
Metalava configuration file. It just covers the name of the surface
and whether it extends another surface. Future changes will add
information about the criteria, e.g. annotations, that are used to
select the contents of the API surface.

Bug: 391554590
Test: m checkapi
Change-Id: Idcd7f6e7d2f02d7cea89a4513ffac89e5dd2dbdf