Add an environment variable for soong-only builds
Some tools/scripts on CI and locally don't accept --soong-only.
To make it easier to disable soong-only for those tools, and to
make it easier for users to disable soong-only for their whole shell,
add an enviornment variable that controls it.

The precedence, from lowest to highest, is: flag, env var, product var.

Bug: 391925617
Test: Manually
Change-Id: Ic828f339544429ac22c56ab31c994a715753caa9