Move required out of common properties
Properties in commonProperties are inherited when creating a new module:
https://cs.android.com/android/platform/superproject/main/+/main:build/soong/android/hooks.go;l=144;drc=b0aabb1f6559d6ae439adeb0691a0124527536cb

This is problematic with required and the autogenerated rro modules.
The autogenerated rro modules are created from an app module, so it
will copy the app's required modules. The autogenerated rro modules
are installed to different partitions from the origional app, so if
you have an app that depends on another app, the second app will not be
installed into two different partitions because we filter by partition,
but the odex/vdex files will be installed from two different sources
to the system_other image and conflict.

Bug: 392702675
Test: m nothing on the failing product
Change-Id: Icbe228a92187bea850908b79571b209475ede77a