cc/rust: Prevent duplicate obj files in staticlibs
When cc_library takes generated sources from a rust_bindgen to produce
the staticfns function, it ends up with duplicate .o files in the
archive. This leads to duplicate symbol collisions against itself if
included whole.

This CL makes sure that the .rs file that the rust_bindgen also produces
is properly ignored and doesn't influence the number of expected .o
files.

Test: m <rust_bindgen w/ static_inline_library linked whole>
Change-Id: I5c7d0db0fc5e2a825f468348b4a709f94a374782