Disable target_files.zip generation in soong-only builds
The current implementation chains the non-hermetic and hermetic file
creation in a single cmd. This causes `m droid` in --soong-only to do
some redundant work. Since the hermetic file is used in
target_files.zip, and this file is currently not fully migrated to soong
builds, comment this out for now.

Test: Set PRODUCT_SOONG_ONLY locally
Test: m droid
Bug: 393203512
Change-Id: Iff3a6a553922533c58b3d0206898bd627281c3fb