Create install rule for loadplugins without adding to blueprint_tools
m checkbuild --soong-only currently runs into errors for "Test for
dangling rules" since the install rule for loadplugins is not created in
`m --soong-only`. In regular `m` builds kati generates the installation
rule since `LOCAL_SOONG_INSTALLED_MODULE` for `loadplugins` is empty in
the generated Android.mk file.

To solve the missing rule in soong-only mode, always generate the
installation rule.

To prevent repeated reanalysis in `m`, do not add the installed file to
`blueprint_tools`, which is a transitive dep of droid via droid_targets.

Bug: 394096939
Test: m blueprint_tools && m blueprint_tools (second one is noop)
Test: m checkbuild --soong-only (passes at top of stack)
Change-Id: I202bb2fa31543f8e3acbe11a4732f6236cd3b71f