Remove configurable values from providers
Providers are data exported from one module to another.

Configurable properties are the collection of all possible values of a
property, to be resolved under the module's configuration.

If you export a configurable from one module to another and then resolve
it there, it will be evaluated with the wrong configuration.

Bug: 394967253
Test: m nothing
Change-Id: I934b08538bcc3267b3ec44925591f16a1c59292c