Use ld-library-path to point to the standalone libs when enabled
A follow of aosp/3459827 to implement go/standalone-native-device-tests

When `standalone_test` is set to `true` for `cc_test`, use
`ld-library-path` to point to the `standalone-libs` directory to use
those dependencies.

This is to add the `ld-library-path` option within the autogenerated
AndroidTest.xml file. Predefined test config needs to be manually
updated, e.g.:

```
...
<test class="com.android.tradefed.testtype.GTest" >
        <option name="ld-library-path" value="/data/local/tmp/libunwindstack_unit_test/x86_64/standalone-libs:/<some_other_path>" />
...
```

Test: m libunwindstack_unit_test
Change-Id: I41222f1651b746ae040096555369822041980651