Make soong-only builds adb-sync compatible
Historically, the build has put the staging directories for partitions
in out/target/product. But in soong-only builds, the main staging
directory is in the out/soong/.intermediates path for the filesystem
modules, and then later also replicated in out/target/product for
backwards compatibility with tools like adb sync.

However, adb sync determines which files to push by if they have
a different timestamp than on device. Since the device images were
built using a different staging directory than the one adb sync is
looking at, the have different timestamps, and adb sync syncs all files
on the device the first time it's run.

To work around this, ensure the files in both staging directories
have the same timestamp by preserving the timestamp from their
source file with the `cp -p` flag.

This cl only does that change for real files right now, not symlinks.
We could also do it for symlinks in the future, but they're more
complicated and I didn't want to deal with that right now. This means
the first adb sync will resync all symlinks on the device, but there
are relatively few of them and they're small so it's not that big of
an issue.

Long term, we should rewrite adb sync to not depend on staging
directories. If it looks at the actual module intermediates files,
it will see a more consistent timestamp.

Bug: 396466053
Test: Manually ran m sync workflows. Also testing togging soong-only on and off and rebuilding
Change-Id: Ia0076a2f8e58404c5c3a0bf0d33b09b6846e95bd