Finish ninja action statuses when ninja exits
All status data on pending actions is being lost when ninja is killed.
After shutting down the NinjaReader goroutine send FinishAction status
updates for any actions that were running.  This will cause the actions
to be logged to verbose.log.gz and to build.trace.gz when ninja is killed
or when the build is killed with ctrl-C.

Bug: 396440826
Test: examine build.trace.gz and verbose.log.gz after killing ninja
Change-Id: I9cf5b68aeefdace72e8cd7991d730d58f153f323