Reapply "Make fsverity files able to be built independently"
The difference from the original cl is an extra mkdir -p before
copying the apks to their staging dir path. The parent directories
weren't being created before.

Right now, fsverity files are built when building the staging directory.
This is slow, before this cl it takes over a minute to generate the
staging directory, after it's only ~10 seconds.

Fsverity files are now generated before the staging directory is built,
in independant ninja actions.

Bug: 394404628
Test: m aosp_shared_system_image, verified it didn't change before/after
Change-Id: Icf48f0859e6d66ba5ce68d77a7254218d271d526