llndk stub of libbinder_ndk reads systemapi annotation
LLNDK libraries provide stable APIs to both product and vendor
partitions. But we have an exception for libbinder_ndk because of the
Treble boundary.
libbinder_ndk provides different APIs between product and vendor
partitions, because one is in the framework side, and the other is in
the vendor side of the Treble boundary.
libbinder_ndk used systemapi annotation for the product APIs instead
of llndk annotation.
Read the APIs with systemapi annotation to generate the product
variant of the libbinder_ndk llndk stub library.

Bug: 395576642
Test: m libbinder_ndk.product
Change-Id: I66a5b023f0c482b4bbf1da281a88073d1ed4620f