Use built files instead of installed files as deps
When building test zips. Some other rules, like target_files.zip,
are not hermetic, and pick up all files in the installation staging
directories. This test_package module was causing more things to be
placed into the staging directory and thus packaged into
target_files.zip. Copy the built file instead of the installed file
so that the staging directories are not affected.

Bug: 388850000
Test: diff'd soong-built platform_tests.zip before/after this change, no differences
Change-Id: I0e354cd62277aef7f4401d7e2b1d42267d257112