Collect symbols info of jni libs of java modules
symbols info are informations needed to copy the symbols files to the
correct location within $PRODUCT_OUT/symbols directory. This change adds
the logic to collect these information for the jni libs of java modules.
Note that the actual copying of the symbols files will be done in a follow up change.

Test: CI
Bug: 395989947
Change-Id: I23d7006e2c6b18f1193bb601ab102e5c514781a5