Create a validation action to assert `apps` matches prebuilt apex contents
apkcerts.txt built by Make currently contains an entry app in the tree,
but is being migrated to be limited to the list of installed apps for
the device. This unfortunately breaks signing for prebuilt_apex with
apks, since it silently relied on an entry being created by the source
`android_app` module.

https://r.android.com/3552560 introduces a new `apps` property that will
be used to declare this dependency between prebuilt_apex and the cert of
the android app.

This CL introduces a validation action to ensure that the `apps`
property is uptodate. It uses apex-ls to get the actual list of
installed apks, and compares it against the contents of the AppInfo
providers gathered from `apps` property. To support development in
minimal branches like art which might not contain the source app, the
validation action will be triggered when building apkcerts.txt (either
in Soong or Make), and will not be in the transitive closure of `m
$my_prebuilt_apex`.

Test: m
Bug: 399788149
Change-Id: Idb8ec3ca3bbeee9bd4c09a7d5cd981cc35d765dc