Disallow cross partition overrides in android_device
Make supports cross partition overrides, but this functionality will not
be supported in Soong due to the technical complexities required in
querying/providing this information to the filesystem dependencies. Add
a check on `android_device` to prevent this unsupported behavior.

Long term when we have `android_device` and `android_fileystem` modules
checked into the tree, `android_filesystem_defaults` might be a good
solution if a cross partition override is required.

Bug: 375035189
Test: m nothing --no-skip-soong-tests
Change-Id: Ibd2c32dffc4fed5a50fff233a3b86e87901d534c